+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## 0.10.0 - 2019-01-30
+## 1.x.x - 2019-##-
+
+1.
+
+## 0.10.0 - 2019-04-01
 
 1. Refactor to use new Api search types like `SearchOptionsV2`. See D24185
 2. Mark old `SearchOption` types as Deprecated or Old.